refactor: remove CodeReader module, distribute functions to SDK modules
- Move parseFile, isInherited, ObjectFromCode to TypesSDK as SwiftParser
- Move linesOfCode to LOCSDK as ClocRunner
- Update Package.swift to remove CodeReader target
- Move tests from CodeReaderTests to TypesSDKTests
- Remove Sources/CodeReader and Tests/CodeReaderTests directories

Closes #48
